:truck: Move: 프로젝트 페이지 컴포넌트 파일 정리
Move: 프로젝트 페이지 컴포넌트 파일 정리
